--- data/sleep/api/src/main/kotlin/com/daangn/data/sleep/api/SleepRepository.kt
@@ -1,11 +1,0 @@
-package com.daangn.data.sleep.api
-
-import kotlinx.coroutines.flow.Flow
-
-interface SleepRepository {
-
-    suspend fun setSleep(isSleep: Boolean)
-
-    fun getSleepFlow(): Flow<Boolean>
-
-}
