+++ data/sleep/api/src/main/kotlin/com/titi/data/sleep/api/SleepRepository.kt
@@ -1,0 +1,11 @@
+package com.titi.data.sleep.api
+
+import kotlinx.coroutines.flow.Flow
+
+interface SleepRepository {
+
+    suspend fun setSleep(isSleep: Boolean)
+
+    fun getSleepFlow(): Flow<Boolean>
+
+}
