--- data/sleep/impl/src/main/kotlin/com/daangn/data/sleep/impl/di/DataStoreModule.kt
@@ -1,22 +1,0 @@
-package com.daangn.data.sleep.impl.di
-
-import android.content.Context
-import com.daangn.data.sleep.impl.local.SleepDataStore
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal object DataStoreModule {
-
-    @Provides
-    @Singleton
-    fun provideSleepDataStore(
-        @ApplicationContext context: Context
-    ) = SleepDataStore(context)
-
-}
