--- data/sleep/impl/src/main/kotlin/com/daangn/data/sleep/impl/di/RepositoryModule.kt
@@ -1,21 +1,0 @@
-package com.daangn.data.sleep.impl.di
-
-import com.daangn.data.sleep.api.SleepRepository
-import com.daangn.data.sleep.impl.repository.SleepRepositoryImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface RepositoryModule {
-
-    @Binds
-    @Singleton
-    fun provideSleepRepository(
-        sleepRepositoryImpl: SleepRepositoryImpl
-    ): SleepRepository
-
-}
