--- data/sleep/impl/src/main/kotlin/com/daangn/data/sleep/impl/local/SleepDataStore.kt
@@ -1,29 +1,0 @@
-package com.daangn.data.sleep.impl.local
-
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.booleanPreferencesKey
-import androidx.datastore.preferences.preferencesDataStore
-import com.titi.core.util.readFlowValue
-import com.titi.core.util.storeValue
-
-internal class SleepDataStore(context: Context) {
-
-    private val dataStore: DataStore<Preferences> = context.dataStore
-
-    suspend fun setSleep(isSleep: Boolean) {
-        dataStore.storeValue(SLEEP_KEY, isSleep)
-    }
-
-    fun getSleepFlow() = dataStore.readFlowValue(SLEEP_KEY)
-
-    companion object {
-
-        private const val SLEEP_PREF_NAME = "sleepPrefName"
-        private val SLEEP_KEY = booleanPreferencesKey("sleepKey")
-
-        private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = SLEEP_PREF_NAME)
-    }
-
-}
